% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{buildPlaylistDF}
\alias{buildPlaylistDF}
\title{Bilds the dataframe.
 Takes the initial query as an argument}
\usage{
buildPlaylistDF(query)
}
\description{
Bilds the dataframe.
 Takes the initial query as an argument
}
