% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cuckoo}
\alias{cuckoo}
\title{Counts of common cuckoo from the Swedish Bird Survey}
\format{
\subsection{\code{cuckoo}}{

A data frame with 6 columns.
\describe{
\item{count}{Number of individuals counted.}
\item{yr}{Year}
\item{route}{Id of the survey route.}
\item{county}{County where route is located.}
\item{lon}{Longitude}
\item{lon}{Latitude}
}
}
}
\source{
\doi{10.15468/hd6w0r}
}
\usage{
cuckoo
}
\description{
Counts per survey route and year between 2000 and 2021 of common cuckoos from the
Swedish Bird Survey.
}
\keyword{datasets}
