% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackClass.R
\name{featurePlot}
\alias{featurePlot}
\title{TODO: plotting method for audio features. Requires new getter that constructs
a list of lists of audio features from playlists that we want to explore.
Polots then the selected playlists on 2d grapgh that has x and y axes that
represent two of the features.
Plot audio features.}
\usage{
featurePlot(x, y, trakcs)
}
\arguments{
\item{x}{Name of the feature to be plotted on x-axis.}

\item{y}{Naem of the feature to be plotted on y-axis.}

\item{tracks}{A list of audioFeatures objects.}
}
\description{
Plots one feature against other given a list of audio feature objects.
}
