% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceess.R
\name{getAccesstoken}
\alias{getAccesstoken}
\title{Sends POST request to Spotify API and uses the authenthication ID and the
Secret key
provided with setAuthentication function}
\usage{
getAccesstoken()
}
\description{
Sends POST request to Spotify API and uses the authenthication ID and the
Secret key
provided with setAuthentication function
}
