% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist.R
\name{getArtist}
\alias{getArtist}
\title{TODO Functions that allow pulling data on artists
https://developer.spotify.com/documentation/web-api/reference-beta/#category-artists
Info on artist.
returns a list that contains the info on the artist}
\usage{
getArtist(artistID)
}
\arguments{
\item{artistID}{spotify ID of the artist.}
}
\description{
TODO Functions that allow pulling data on artists
https://developer.spotify.com/documentation/web-api/reference-beta/#category-artists
Info on artist.
returns a list that contains the info on the artist
}
