% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist.R
\name{getArtistTopTracks}
\alias{getArtistTopTracks}
\title{Returns the toptracks of artist.}
\usage{
getArtistTopTracks(artistID, country)
}
\arguments{
\item{artistID}{The Spotify ID of the artist}

\item{country}{An ISO 3166-1 alpha-2 country code or the string}
}
\description{
Returns the toptracks of artist.
}
