% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{getAudioAnalysis}
\alias{getAudioAnalysis}
\title{Returns audio analysisi object.}
\usage{
getAudioAnalysis(trackID)
}
\arguments{
\item{trackID}{Spotify ID}
}
\value{
An audio object
}
\description{
Returns audio analysisi object.
}
