% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{getAudioFeatures}
\alias{getAudioFeatures}
\title{TODO: Change this so that there is no 100 song limit.
Get audio features for track.}
\usage{
getAudioFeatures(IDs)
}
\arguments{
\item{IDs}{a vector of IDs. Maximum lengths is 100 IDs.}
}
\description{
TODO: Change this so that there is no 100 song limit.
Get audio features for track.
}
