% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{getTrack}
\alias{getTrack}
\title{A function that returns info on a track based on its SpotiyID.}
\usage{
getTrack(trackID)
}
\arguments{
\item{trackID}{Tjhe Spotify ID of the track}
}
\value{
A list that contains the track name as chr, track id as chr,
popularity as numeric and possible genres as an vector. It should be noted
that the genre comes from artist query and might not entirely mach the genre
of the song.
}
\description{
A function that returns info on a track based on its SpotiyID.
}
