% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackClass.R
\name{plot.audio}
\alias{plot.audio}
\title{Plots audio object.}
\usage{
plot.audio(x, timeRange = NULL, showSegments = FALSE)
}
\arguments{
\item{timeRange}{A vector of length 2 that contains the start and finnish of
range you desire to explore.}

\item{shoSegments}{Whether to show segments of not, since they are numerous so they might
get on the way of seeing other stuff.}
}
\description{
Plots audio object.
}
