% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/9_Functions_test_package_4Methods_random_XYsites.R
\name{plot.power}
\alias{plot.power}
\title{Plot results of power analysis}
\usage{
plot.power(pwr, n.species, nsims, effect.size, n.park, species.list, n.cores)
}
\arguments{
\item{pwr}{An array containing the results of the power analysis}

\item{n.species}{The number of species in which to estimate power}

\item{nsims}{The number of simulations across each core}

\item{effect.size}{A vector specifying the proportional decline in occupancy}

\item{n.park}{The number of nested management units in which to estimate power}

\item{species.list}{A vector specifying the name of each species analysed}

\item{n.cores}{A matrix specifying the name of each species analysed and the relevant detection method}
}
\description{
This function plots results of the power analysis. It first unpacks the list generated by the foreach function and estimates power given the number of simulations 
and the number of cores used. Power is plotted for each species on a single figure. Users can then scroll through separate plots for power at a park level if it park.power = TRUE
}
