% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceess.R
\name{setAuthentication}
\alias{setAuthentication}
\title{Sets authentication parameters.}
\usage{
setAuthentication(clientID, ClientSecret)
}
\arguments{
\item{clientID}{The client ID token provided by Spotify}

\item{ClientSecret}{The Client Secret token provided by Spotify}
}
\value{
Does not return a value
}
\description{
Sets the authentication id and secret token. Does not return any value,
and is used only to set these parametses as enviromental variables.
}
\author{
Topias Karjalainen
}
