% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/9_Functions_test_package_4Methods_random_XYsites.R
\name{sig.test}
\alias{sig.test}
\title{Significance level function}
\usage{
sig.test(two.tailed, alpha)
}
\arguments{
\item{two.tailed}{Set to TRUE for a two-tailed test, FALSE for a one-tailed test.}

\item{alpha}{Set the type one error rate. Choose between 0.1, 0.05 and 0.01}
}
\description{
This function returns a critical value needed to calculate confidence intervals around the trend parameter given the Type I error rate and whether it is a one or two-tailed significance test.
}
\examples{
#Return the critical value for a two-tailed test with a significance level of 0.05
two.tailed <- TRUE
alpha <- 0.05
sig.test(two.tailed, alpha)

#Return the critical value for a one-tailed test with a significance level of 0.1
two.tailed <- FALSE
alpha <- 0.1
sig.test(two.tailed, alpha)
}
