% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotr-package.R
\docType{package}
\name{spotr-package}
\alias{spotr}
\alias{spotr-package}
\title{spotr: Estimate Spatial Population Indices from Ecological Abundance Data}
\description{
Compute relative or absolute population trends across space using predictions from models fitted to ecological population abundance data. The package supports models fitted by 'mgcv' and 'brms', and draws from posterior predictive distributions.
}
\author{
\strong{Maintainer}: Jonas Knape \email{jonas.knape@slu.se} (\href{https://orcid.org/0000-0002-8012-5131}{ORCID})

}
\keyword{internal}
