% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/9_Functions_test_package_4Methods_random_XYsites.R
\name{stochastic.disturbance}
\alias{stochastic.disturbance}
\title{Stochastic disturbance function}
\usage{
stochastic.disturbance(time.hist, Pr.dist, n.sites)
}
\arguments{
\item{time.hist}{An array specifying the time since a disturbance at each monitoring site prior to simulations}

\item{Pr.dist}{A vector specifying the probability of a cell being disturbed given the time since the last disturbance.}

\item{n.sites}{The number of sites monitored during each simulation}
}
\description{
This function models the incidence of a stochastic disturbance at monitoring sites given the probability of a cell being disturbed as a function of time since the last disturbance.
It returns a vector specifying whether each site is disturbed (1) or not (0) in a given year jj
}
