% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swe_map}
\alias{swe_map}
\title{Map of Swedish counties}
\format{
\subsection{\code{cuckoo}}{

An sf data frame with 3 columns.
\describe{
\item{county}{Name of county.}
\item{area}{Area of county in km^2.}
\item{geometry}{sf geometry column}
}
}
}
\source{
\url{https://www.scb.se/hitta-statistik/regional-statistik-och-kartor/regionala-indelningar/digitala-granser/}
}
\usage{
swe_map
}
\description{
Polygons of Swedish counties and their areas.
}
\keyword{datasets}
