% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{buildPlaylistDF}
\alias{buildPlaylistDF}
\title{Builds the dataframe.
Takes the initial query as an argument}
\usage{
buildPlaylistDF(query)
}
\description{
Builds the dataframe.
Takes the initial query as an argument
}
