% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{buildTrackDF}
\alias{buildTrackDF}
\title{Hidden function. Builds the data frames that contain the tracks.}
\usage{
buildTrackDF(playlistURL)
}
\arguments{
\item{playlistURL}{An URL}
}
\description{
Hidden function. Builds the data frames that contain the tracks.
}
