% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackClass.R
\name{featurePlot}
\alias{featurePlot}
\title{TODO: Add legends to the  plots made with this function.
TODO: A function that creates ready lists of playlists to be plotted.
Plot audio features.}
\usage{
featurePlot(x, y, trakcs)
}
\arguments{
\item{x}{Name of the feature to be plotted on x-axis.}

\item{y}{Naem of the feature to be plotted on y-axis.}

\item{tracks}{A list of audioFeatures objects, or a list of lists(different
playlists perhaps).}
}
\description{
Plots one feature against other given a list of audio feature objects.
}
