% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{getPlaylistTracks}
\alias{getPlaylistTracks}
\title{Creates a list of data.frames that contain playlist tarcks.}
\usage{
getPlaylistTracks(playlists)
}
\arguments{
\item{playlists}{The data.frame returned by getUserPlaylists function}
}
\description{
Creates a list of data.frames that contain playlist tarcks.
}
