% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{getTrack}
\alias{getTrack}
\title{A function that returns info on a track based on its SpotiyID.}
\usage{
getTrack(trackID)
}
\arguments{
\item{trackID}{Tjhe Spotify ID of the track}
}
\description{
A function that returns info on a track based on its SpotiyID.
}
