% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{getUserPlayslists}
\alias{getUserPlayslists}
\title{Retuns a data frame of the users playlists.}
\usage{
getUserPlayslists(userID)
}
\arguments{
\item{userID}{The wanted user id as a string}
}
\description{
Retuns a data frame of the users playlists.
}
